+++ src/Services/Basket/BasketAPI/Models/ShoppingCarItem.cs
@@ -6,8 +6,8 @@
         public string ProductName { get; set; } = default!;
         public int Quantity { get; set; } = default!;
         public string Color { get; set; } = default!;
-        public double Price { get; set; } = default!;
-        public double DiscountValue { get; set; } = 0;
+        public decimal Price { get; set; } = default!;
+        public decimal DiscountValue { get; set; } = 0;
         public Guid ProductId { get; set; } = default !;
     }
 }
